pyproject.toml
[project]
name = "yolo-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "streamlit>=1.48.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.178",
]
